Midification
- splitted the data parsing for better readidibility
- moved and changed 'default' from parse_data to const for better config
