+++ custom_components/sonarr_upcoming_media/const.py
@@ -5,3 +5,13 @@ DOMAIN: Final = "sonarr_upcoming_media"
 CONF_DAYS: Final = 'days'
 CONF_URLBASE: Final = 'urlbase'
 CONF_MAX: Final = 'max'
+
+
+DEFAULT_PARSE_DICT: Final = {
+    'title_default': '$title',
+    'line1_default': '$episode',
+    'line2_default': '$release',
+    'line3_default': '$rating - $runtime',
+    'line4_default': '$number - $studio',
+    'icon': 'mdi:arrow-down-bold'
+}
